requirements.txt
greenlet==3.3.0
numpy==2.2.6
pandas==2.3.3
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.45
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
